Refactoring.  Extracting AfterMidnight module.
